requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-autoslug==1.9.8
django-cities-light==3.8.1
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
phonenumbers==8.12.31
Pillow==8.3.1
progressbar2==3.53.2
psycopg2==2.9.1
pycodestyle==2.7.0
python-dotenv==0.19.0
python-utils==2.5.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.3.1
